feat: add home page
